+++ src/services/hid/Composition.test.ts
@@ -5,11 +5,10 @@ import {
   MOD_ALT,
   MOD_CTL,
   MOD_GUI,
-  MOD_LEFT,
-  MOD_RIGHT,
   MOD_SFT,
 } from './Composition';
 import { OP_SH_TAP_TOGGLE } from './compositions/SwapHandsComposition';
+import { MOD_LEFT, MOD_RIGHT } from './Constraints';
 
 const EXPECT_BASIC_LIST = [0b0000_0000_0000_0000, 0b0000_0000_1111_1111];
 const EXPECT_MODS_LIST = [0b0000_0001_0000_0000, 0b0001_1111_1111_1111];
