+++ src/store/reducers.ts
@@ -1,50 +1,50 @@
 import immer from 'immer';
 import { WritableDraft } from 'immer/dist/internal';
 import {
+  ANYKEYCODEKEY_ACTIONS,
+  ANYKEYCODEKEY_ADD_ANYKEY,
+  ANYKEYCODEKEY_UPDATE_ANYKEY,
+  APP_ACTIONS,
+  APP_PACKAGE_INIT,
+  APP_REMAPS_CLEAR,
+  APP_REMAPS_INIT,
+  APP_REMAPS_REMOVE_KEY,
+  APP_REMAPS_SET_KEY,
+  APP_REMAPS_SET_KEYS,
+  APP_TEST_MATRIX_UPDATE,
+  APP_TESTED_MATRIX_CLEAR,
+  APP_UPDATE_KEYBOARD_SIZE,
+  APP_UPDATE_LANG_LABEL,
+  APP_UPDATE_SETUP_PHASE,
+  APP_UPDATE_SIGNED_IN,
+  HEADER_ACTIONS,
+  HEADER_UPDATE_FLASHING,
   KEYCODEKEY_ACTIONS,
+  KEYCODEKEY_CLEAR,
+  KEYCODEKEY_UPDATE_DRAGGING_KEY,
   KEYCODEKEY_UPDATE_HOVER_KEY,
   KEYCODEKEY_UPDATE_SELECTED_KEY,
   KEYCODES_ACTIONS,
   KEYCODES_UPDATE_MACRO,
-  KEYMAP_UPDATE_SELECTED_LAYER,
-  KEYMAP_ACTIONS,
-  NOTIFICATION_ACTIONS,
-  NOTIFICATION_ADD_ERROR,
-  NOTIFICATION_ADD_WARN,
-  HEADER_UPDATE_FLASHING,
-  HEADER_ACTIONS,
-  KEYCODEKEY_UPDATE_DRAGGING_KEY,
   KEYDIFF_ACTIONS,
-  KEYDIFF_UPDATE_KEYDIFF,
   KEYDIFF_CLEAR_KEYDIFF,
-  KEYMAP_UPDATE_SELECTED_POS,
+  KEYDIFF_UPDATE_KEYDIFF,
+  KEYMAP_ACTIONS,
   KEYMAP_CLEAR_SELECTED_POS,
-  APP_ACTIONS,
-  APP_REMAPS_SET_KEY,
-  APP_REMAPS_INIT,
-  APP_REMAPS_REMOVE_KEY,
-  APP_PACKAGE_INIT,
-  NOTIFICATION_REMOVE,
-  NOTIFICATION_ADD_INFO,
-  NOTIFICATION_ADD_SUCCESS,
-  APP_UPDATE_SETUP_PHASE,
-  APP_REMAPS_CLEAR,
-  KEYCODEKEY_CLEAR,
-  ANYKEYCODEKEY_ACTIONS,
-  ANYKEYCODEKEY_ADD_ANYKEY,
-  ANYKEYCODEKEY_UPDATE_ANYKEY,
-  LAYOUT_OPTIONS_ACTIONS,
-  LAYOUT_OPTIONS_UPDATE_SELECTED_OPTION,
-  LAYOUT_OPTIONS_INIT_SELECTED_OPTION,
-  APP_UPDATE_KEYBOARD_SIZE,
-  APP_UPDATE_LANG_LABEL,
-  APP_UPDATE_SIGNED_IN,
-  APP_REMAPS_SET_KEYS,
-  LAYOUT_OPTIONS_RESTORE_SELECTED_OPTIONS,
   KEYMAP_TOOLBAR_ACTIONS,
   KEYMAP_TOOLBAR_TEST_MATRIX_MODE,
-  APP_TESTED_MATRIX_CLEAR,
-  APP_TEST_MATRIX_UPDATE,
+  KEYMAP_UPDATE_SELECTED_LAYER,
+  KEYMAP_UPDATE_SELECTED_POS,
+  LAYOUT_OPTIONS_ACTIONS,
+  LAYOUT_OPTIONS_INIT_SELECTED_OPTION,
+  LAYOUT_OPTIONS_RESTORE_SELECTED_OPTIONS,
+  LAYOUT_OPTIONS_UPDATE_SELECTED_OPTION,
+  NOTIFICATION_ACTIONS,
+  NOTIFICATION_ADD_ERROR,
+  NOTIFICATION_ADD_INFO,
+  NOTIFICATION_ADD_SUCCESS,
+  NOTIFICATION_ADD_WARN,
+  NOTIFICATION_REMOVE,
 } from '../actions/actions';
 import {
   HID_ACTIONS,
@@ -91,12 +91,16 @@ import {
   KEYBOARDS_CREATE_DEFINITION_ACTIONS,
   KEYBOARDS_CREATE_DEFINITION_CLEAR,
   KEYBOARDS_CREATE_DEFINITION_UPDATE_AGREEMENT,
+  KEYBOARDS_CREATE_DEFINITION_UPDATE_AUTHOR_TYPE,
+  KEYBOARDS_CREATE_DEFINITION_UPDATE_CONTACT_INFORMATION,
   KEYBOARDS_CREATE_DEFINITION_UPDATE_FIRMWARE_CODE_PLACE,
   KEYBOARDS_CREATE_DEFINITION_UPDATE_FORKED_REPOSITORY_EVIDENCE,
   KEYBOARDS_CREATE_DEFINITION_UPDATE_FORKED_REPOSITORY_URL,
   KEYBOARDS_CREATE_DEFINITION_UPDATE_JSON_FILENAME,
   KEYBOARDS_CREATE_DEFINITION_UPDATE_JSON_STRING,
   KEYBOARDS_CREATE_DEFINITION_UPDATE_KEYBOARD_DEFINITION,
+  KEYBOARDS_CREATE_DEFINITION_UPDATE_ORGANIZATION_EVIDENCE,
+  KEYBOARDS_CREATE_DEFINITION_UPDATE_ORGANIZATION_ID,
   KEYBOARDS_CREATE_DEFINITION_UPDATE_OTHER_PLACE_HOW_TO_GET,
   KEYBOARDS_CREATE_DEFINITION_UPDATE_OTHER_PLACE_PUBLISHER_EVIDENCE,
   KEYBOARDS_CREATE_DEFINITION_UPDATE_OTHER_PLACE_SOURCE_CODE_EVIDENCE,
@@ -105,47 +109,42 @@ import {
   KEYBOARDS_EDIT_DEFINITION_ACTIONS,
   KEYBOARDS_EDIT_DEFINITION_ADD_ADDITIONAL_DESCRIPTIONS,
   KEYBOARDS_EDIT_DEFINITION_CLEAR,
+  KEYBOARDS_EDIT_DEFINITION_CLEAR_FIRMWARE_FORM,
   KEYBOARDS_EDIT_DEFINITION_DELETE_ADDITIONAL_DESCRIPTIONS,
   KEYBOARDS_EDIT_DEFINITION_INIT,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_ADDITIONAL_DESCRIPTIONS,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_AGREEMENT,
+  KEYBOARDS_EDIT_DEFINITION_UPDATE_AUTHOR_TYPE,
+  KEYBOARDS_EDIT_DEFINITION_UPDATE_CONTACT_INFORMATION,
+  KEYBOARDS_EDIT_DEFINITION_UPDATE_DEFAULT_BOOTLOADER_TYPE,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_DESCRIPTION,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_FEATURE,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_FEATURES,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_FIRMWARE_CODE_PLACE,
+  KEYBOARDS_EDIT_DEFINITION_UPDATE_FIRMWARE_DESCRIPTION,
+  KEYBOARDS_EDIT_DEFINITION_UPDATE_FIRMWARE_FILE,
+  KEYBOARDS_EDIT_DEFINITION_UPDATE_FIRMWARE_NAME,
+  KEYBOARDS_EDIT_DEFINITION_UPDATE_FIRMWARE_SOURCE_CODE_URL,
+  KEYBOARDS_EDIT_DEFINITION_UPDATE_FLASH_SUPPORT,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_FORKED_REPOSITORY_EVIDENCE,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_FORKED_REPOSITORY_URL,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_JSON_FILENAME,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_JSON_STRING,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_KEYBOARD_DEFINITION,
+  KEYBOARDS_EDIT_DEFINITION_UPDATE_MAIN_IMAGE_UPLOADED_RATE,
+  KEYBOARDS_EDIT_DEFINITION_UPDATE_MAIN_IMAGE_UPLOADING,
+  KEYBOARDS_EDIT_DEFINITION_UPDATE_ORGANIZATION_EVIDENCE,
+  KEYBOARDS_EDIT_DEFINITION_UPDATE_ORGANIZATION_ID,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_OTHER_PLACE_HOW_TO_GET,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_OTHER_PLACE_PUBLISHER_EVIDENCE,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_OTHER_PLACE_SOURCE_CODE_EVIDENCE,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_PRODUCT_NAME,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_QMK_REPOSITORY_FIRST_PULL_REQUEST_URL,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_STORES,
-  KEYBOARDS_EDIT_DEFINITION_UPDATE_MAIN_IMAGE_UPLOADED_RATE,
-  KEYBOARDS_EDIT_DEFINITION_UPDATE_MAIN_IMAGE_UPLOADING,
-  KEYBOARDS_EDIT_DEFINITION_UPDATE_WEBSITE_URL,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_SUB_IMAGE_UPLOADED_RATE,
   KEYBOARDS_EDIT_DEFINITION_UPDATE_SUB_IMAGE_UPLOADING,
-  KEYBOARDS_CREATE_DEFINITION_UPDATE_CONTACT_INFORMATION,
-  KEYBOARDS_EDIT_DEFINITION_UPDATE_CONTACT_INFORMATION,
-  KEYBOARDS_EDIT_DEFINITION_UPDATE_FIRMWARE_NAME,
-  KEYBOARDS_EDIT_DEFINITION_UPDATE_FIRMWARE_FILE,
-  KEYBOARDS_EDIT_DEFINITION_UPDATE_FIRMWARE_DESCRIPTION,
-  KEYBOARDS_EDIT_DEFINITION_CLEAR_FIRMWARE_FORM,
-  KEYBOARDS_EDIT_DEFINITION_UPDATE_FIRMWARE_SOURCE_CODE_URL,
-  KEYBOARDS_EDIT_DEFINITION_UPDATE_FLASH_SUPPORT,
-  KEYBOARDS_CREATE_DEFINITION_UPDATE_AUTHOR_TYPE,
-  KEYBOARDS_CREATE_DEFINITION_UPDATE_ORGANIZATION_ID,
-  KEYBOARDS_CREATE_DEFINITION_UPDATE_ORGANIZATION_EVIDENCE,
-  KEYBOARDS_EDIT_DEFINITION_UPDATE_ORGANIZATION_EVIDENCE,
-  KEYBOARDS_EDIT_DEFINITION_UPDATE_ORGANIZATION_ID,
-  KEYBOARDS_EDIT_DEFINITION_UPDATE_AUTHOR_TYPE,
-  KEYBOARDS_EDIT_DEFINITION_UPDATE_DEFAULT_BOOTLOADER_TYPE,
+  KEYBOARDS_EDIT_DEFINITION_UPDATE_WEBSITE_URL,
 } from '../actions/keyboards.actions';
-import { MOD_LEFT } from '../services/hid/Composition';
 import { LayoutOption } from '../components/configure/keymap/Keymap';
 import {
   CATALOG_APP_ACTIONS,
@@ -195,6 +194,7 @@ import {
   UPLOAD_FIRMWARE_DIALOG_UPDATE_FIRMWARE_FILE_BUFFER,
   UPLOAD_FIRMWARE_DIALOG_UPDATE_OPEN,
 } from '../actions/firmware.action';
+import { MOD_LEFT } from '../services/hid/Constraints';
 
 export type Action = { type: string; value: any };
